requirements.txt
Django==1.8.2
Pillow
MySQL-python==1.2.5
django-crispy-forms==1.4.0
python-dateutil
django-registration-redux==1.3
python-social-auth
django-coverage==1.2.4
coverage==3.7.1